backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
google-auth>=2.27.0
python-dotenv>=1.0.1
pydantic[email]>=2.6.0
openai>=1.10.0
websockets>=12.0
requests>=2.31.0
aiosmtplib>=3.0.0
email-templates>=0.2.2
jinja2>=3.1.0